moved save function and state to editor.js removed keyword function that was crashing the app
