+++ client/src/App.js
@@ -10,6 +10,7 @@ import TopNav from './components/TopNav';
 import { Container } from 'reactstrap';
 import UserContext from './utils/UserContext';
 import NewSnip from './pages/newsnip'
+
 const App = () => {
   const [userData, setUserData] = useState({
     firstname: '',
@@ -19,187 +20,150 @@ const App = () => {
     password: '',
   });
 
-  const [snipData, setSnipData] = useState({
-    title: '',
-    userId: '',
-    public: '',
-    scriptType: '',
-    snip: '',
-    scriptTypeTwo: '',
-    snipTwo: '',
-    scriptTypeThree: '',
-    snipThree: '',
-    scripTypeThree: '',
-    keywords: '',
-    dateCreated: '',
-    comments: '',
-  });
+ 
+  const [loggedIn, setLoggedin] = useState(false);
+  const [user, setUser] = useState(null);
+  const [failureMessage, setFailureMessage] = useState(null);
 
-  const saveFromEditor = (event) => {
+  useEffect(() => {
+    isLoggedIn();
+  }, []);
+
+  const handleInputChange = (event) => {
+    const { name, value } = event.target;
+    setUserData({ ...userData, [name]: value });
+  };
+
+  const handleLogin = (event) => {
+    event.preventDefault();
+    const data = {
+      username: userData.username,
+      password: userData.password,
+    };
+    if (userData.username && userData.password) {
+      API.login(data)
+        .then((user) => {
+          if (user.data.loggedIn) {
+            setUser(user.data.user);
+            setLoggedin(true)
+
+            console.log('log in successful');
+            window.location.href = '/profile';
+          } else {
+            console.log('Something went wrong :(');
+            alert('Login failed, Please try again.');
+          }
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    }
+  };
+
+  const handleSignup = (event) => {
     event.preventDefault();
     try {
-      const codeData = {
-        title: snipData.title,
-        userId: snipData.userId,
-        public: snipData.public,
-        scriptType: snipData.codeType,
-        snip: snipData.snip,
-        scriptTypeTwo: snipData.scriptTypeTwo,
-        snipTwo: snipData.scriptTypeTwo,
-        scriptTypeThree: snipData.scriptTypeTwo,
-        snipThree: snipData.scriptTypeTwo,
-        keywords: snipData.keywords,
-        dateCreated: snipData.dateCreated,
-        comments: snipData.comments
-      };
-      codeAPI.saveSnip(codeData)
-    } catch (error) {
-      console.log('App -> error', error);
-  }
-  }
-      const [loggedIn, setLoggedin] = useState(false);
-      const [user, setUser] = useState(null);
-      const [failureMessage, setFailureMessage] = useState(null);
-
-      useEffect(() => {
-        isLoggedIn();
-      }, []);
-
-      const handleInputChange = (event) => {
-        const { name, value } = event.target;
-        setUserData({ ...userData, [name]: value });
+      const data = {
+        firstname: userData.firstname,
+        lastname: userData.lastname,
+        email: userData.email,
+        username: userData.username,
+        password: userData.password,
       };
 
-      const handleLogin = (event) => {
-        event.preventDefault();
-        const data = {
-          username: userData.username,
-          password: userData.password,
-        };
-        if (userData.username && userData.password) {
-          API.login(data)
-            .then((user) => {
+      if (userData.username && userData.password) {
+        API.signup(data)
+          .then((user) => {
+            if (user.data === 'email is already in use') {
+              alert('Email already in use.');
+            }
+            if (user.data.loggedIn) {
               if (user.data.loggedIn) {
                 setUser(user.data.user);
-                setLoggedin(true)
-
+                setLoggedin(true);
+                console.log(user)
                 console.log('log in successful');
                 window.location.href = '/profile';
               } else {
-                console.log('Something went wrong :(');
-                alert('Login failed, Please try again.');
+                console.log('something went wrong :(');
+                console.log(user.data);
+                setFailureMessage(user.data);
               }
-            })
-            .catch((error) => {
-              console.log(error);
-            });
-        }
-      };
-
-      const handleSignup = (event) => {
-        event.preventDefault();
-        try {
-          const data = {
-            firstname: userData.firstname,
-            lastname: userData.lastname,
-            email: userData.email,
-            username: userData.username,
-            password: userData.password,
-          };
-
-          if (userData.username && userData.password) {
-            API.signup(data)
-              .then((user) => {
-                if (user.data === 'email is already in use') {
-                  alert('Email already in use.');
-                }
-                if (user.data.loggedIn) {
-                  if (user.data.loggedIn) {
-                    setUser(user.data.user);
-                    setLoggedin(true);
-                    console.log(user)
-                    console.log('log in successful');
-                    window.location.href = '/profile';
-                  } else {
-                    console.log('something went wrong :(');
-                    console.log(user.data);
-                    setFailureMessage(user.data);
-                  }
-                }
-              })
-              .catch((error) => {
-                console.log(error);
-              });
-          }
-        } catch (error) {
-          console.log('App -> error', error);
-        }
-      };
-
-      const isLoggedIn = () => {
-        if (!loggedIn) {
-          API.isLoggedIn().then((user) => {
-            if (user.data.loggedIn) {
-              setUser(user.data.user);
-              setLoggedin(true);
-              console.log(user)
-            } else {
-              console.log(user.data.message);
             }
+          })
+          .catch((error) => {
+            console.log(error);
           });
+      }
+    } catch (error) {
+      console.log('App -> error', error);
+    }
+  };
+
+  const isLoggedIn = () => {
+    if (!loggedIn) {
+      API.isLoggedIn().then((user) => {
+        if (user.data.loggedIn) {
+          setUser(user.data.user);
+          setLoggedin(true);
+          console.log(user)
+        } else {
+          console.log(user.data.message);
         }
-      };
+      });
+    }
+  };
 
-      const logout = () => {
-        if (loggedIn) {
-          API.logout().then(() => {
-            console.log('logged out successfully');
-            setLoggedin(false);
-            setUser(null);
-          });
-        }
-      };
+  const logout = () => {
+    if (loggedIn) {
+      API.logout().then(() => {
+        console.log('logged out successfully');
+        setLoggedin(false);
+        setUser(null);
+      });
+    }
+  };
 
-      const contextValue = {
-        userData,
-        loggedIn,
-        user,
-        failureMessage,
-        handleInputChange,
-        handleLogin,
-        handleSignup,
-        logout,
-      };
-      return (
-        <UserContext.Provider value={contextValue}>
-          <Router>
-            <div>
-              <TopNav />
-              <Switch>
-                <Route exact path="/" component={Home} />
-                <Route
-                  exact
-                  path="/login"
-                  render={() => <Auth action="login" />}
-                />
-                <Route
-                  exact
-                  path="/signup"
-                  render={() => <Auth action="signup" />}
-                />
-                <Route
-                  exact
-                  path="/newSnip"
-                  component={NewSnip}
-                  render={() => <Auth action="newSnip" />}
-                />
-                <Route exact path="/profile" component={Profile} />
-                <Route render={NoMatch} />
-              </Switch>
-            </div>
-          </Router>
-        </UserContext.Provider>
-      );
-    };
+  const contextValue = {
+    userData,
+    loggedIn,
+    user,
+    failureMessage,
+    handleInputChange,
+    handleLogin,
+    handleSignup,
+    logout,
+  };
+  return (
+    <UserContext.Provider value={contextValue}>
+      <Router>
+        <div>
+          <TopNav />
+          <Switch>
+            <Route exact path="/" component={Home} />
+            <Route
+              exact
+              path="/login"
+              render={() => <Auth action="login" />}
+            />
+            <Route
+              exact
+              path="/signup"
+              render={() => <Auth action="signup" />}
+            />
+            <Route
+              exact
+              path="/newSnip"
+              component={NewSnip}
+              render={() => <Auth action="newSnip" />}
+            />
+            <Route exact path="/profile" component={Profile} />
+            <Route render={NoMatch} />
+          </Switch>
+        </div>
+      </Router>
+    </UserContext.Provider>
+  );
+};
 
-    export default App;
+export default App;
